fix: zod のバージョンを更新し、設定読み込みロジックとパターンマッチングの詳細を追加
