+++ deno.lock
@@ -40,7 +40,8 @@
     "npm:yaml@*": "2.7.1",
     "npm:zod@*": "3.24.2",
     "npm:zod@3": "3.24.2",
-    "npm:zod@^3.24.2": "3.24.2"
+    "npm:zod@^3.24.2": "3.24.2",
+    "npm:zod@^3.24.3": "3.24.3"
   },
   "jsr": {
     "@luca/esbuild-deno-loader@0.11.1": {
@@ -268,7 +269,7 @@
       "dependencies": [
         "content-type",
         "raw-body",
-        "zod"
+        "zod@3.24.2"
       ]
     },
     "@modelcontextprotocol/server-brave-search@0.6.2": {
@@ -640,7 +641,7 @@
         "form-data-encoder",
         "formdata-node",
         "node-fetch",
-        "zod"
+        "zod@3.24.2"
       ]
     },
     "picocolors@1.1.1": {
@@ -726,6 +727,9 @@
     },
     "zod@3.24.2": {
       "integrity": "sha512-lY7CDW43ECgW9u1TcT3IoXHflywfVqDYze4waEz812jR/bZ8FHDsl7pFQoSZTz5N+2NqRXs8GBwnAwo3ZNxqhQ=="
+    },
+    "zod@3.24.3": {
+      "integrity": "sha512-HhY1oqzWCQWuUqvBFnsyrtZRhyPeR7SUGv+C4+MsisMuVfSPx8HpwWqH8tRahSlt6M3PiFAcoeFhZAqIXTxoSg=="
     }
   },
   "workspace": {
@@ -755,7 +759,8 @@
         "dependencies": [
           "jsr:@std/expect@^1.0.15",
           "npm:gpt-tokenizer@^2.1.1",
-          "npm:js-yaml@^4.1.0"
+          "npm:js-yaml@^4.1.0",
+          "npm:zod@^3.24.3"
         ]
       },
       "packages/processors/openai": {
