+++ docs/project-specs/02.core/path-based-config.md
@@ -159,3 +159,59 @@ path_filters: |
    - パターンキャッシュの活用
    - 最適化されたマッチングアルゴリズム
    - 不要なファイルの早期スキップ
+## 設定読み込みロジック (`loadConfig`)
+
+設定ファイル (`.coderabbitai.yaml`) の読み込みは以下の手順で行われます。
+
+1.  **デフォルト設定**: まず、システム定義のデフォルト設定 (`DEFAULT_CONFIG`) で初期化されます。
+2.  **ファイルアクセス**: 設定ファイルの存在と読み取り権限を確認します。
+    *   ファイルが存在しない、または読み取れない場合、警告ログを出力し、デフォルト設定を使用します。
+3.  **読み込みとパース**: ファイル内容を読み込み、YAML としてパースします。
+4.  **形式検証**: パース結果が有効なオブジェクト形式か検証します。
+    *   無効な形式の場合、警告ログを出力し、デフォルト設定を使用します。
+5.  **マージ**: パースされた設定をデフォルト設定にマージします。設定ファイルの値がデフォルト値を上書きします。
+6.  **エラーハンドリング**: 上記以外のエラーが発生した場合 (例: 予期せぬ I/O エラー)、エラーログを出力し、デフォルト設定を使用します。
+
+## パターンマッチングの詳細
+
+### 具体性の計算 (`calculatePatternSpecificity`)
+
+複数のパターンがファイルにマッチした場合の優先順位付けのため、各 glob パターンには「具体性スコア」が計算されます。スコアが高いほど具体的とみなされます。
+
+スコアリングルール:
+
+-   **基本スコア**: パターンの文字長。長いほど高スコア。
+-   **減点**:
+    -   `**` (ダブルワイルドカード): 1つにつき -3 点。
+    -   `*` (シングルワイルドカード): 1つにつき -2 点。
+-   **加点**:
+    -   `.` (拡張子指定): +5 点。
+    -   `{}` (拡張子グループ): +3 点。
+
+### 指示の取得と結合 (`getInstructionsForFile`)
+
+ファイルパスに適用されるレビュー指示は以下の手順で決定されます。
+
+1.  **設定確認**: `file_path_instructions` が設定されているか確認します。なければ空文字列を返します。
+2.  **パターンマッチング**: 設定内の各 `path` パターンがファイルパスにマッチするかを `matchesGlobPattern` で判定します。
+3.  **具体性計算**: マッチした各パターンの具体性スコアを `calculatePatternSpecificity` で計算します。
+4.  **ソート**: マッチした指示を以下の順でソートします。
+    1.  具体性スコアが高い順 (降順)。
+    2.  具体性が同じ場合は、設定ファイル内での元の記述順 (昇順)。
+5.  **結合**: ソートされた指示の `instructions` 文字列を、改行2つ (`\n\n`) で結合して最終的な指示文字列を生成します。
+6.  **エラーハンドリング**: 処理中にエラーが発生した場合、警告ログを出力し、空文字列を返します。
+
+### Glob から正規表現への変換 (`matchesGlobPattern`)
+
+`matchesGlobPattern` 関数は、glob パターンを正規表現に変換してマッチングを行います。
+
+主な変換ルール:
+
+-   `**` → `.*` (スラッシュを含む任意の0文字以上)
+-   `*` → `[^/]*` (スラッシュを含まない任意の0文字以上)
+-   `?` → `[^/]` (スラッシュを含まない任意の1文字)
+-   `{alt1,alt2}` → `(?:alt1|alt2)` (選択肢)
+-   その他の正規表現メタ文字はエスケープされます。
+-   生成された正規表現は `^...$` で囲まれ、完全一致を要求します。
+
+不正な glob パターン (例: `{` が閉じられていない) が指定された場合、警告ログが出力され、そのパターンはマッチしないものとして扱われます (`false` を返す)。
